Beginning to transition all the way
I personally know that I am doing this in a very weird way. Probably the wrong way. But it works. And I can't think of how else to do it. If anyone else is trying to work on this, I'm genuinely sorry. I know. Believe me. I know how bad this is.